requirements.txt
torch >= 1.0
nltk
pandas
tqdm
pyyaml
pyrouge
